Proper error handling when failure occurs

+++ backend/dockerizer/containerizer/code_processor.py
@@ -3,16 +3,16 @@
 This module contains all the necessary fabric tasks
 needed to build the image
 """
-from fabric.api import run,env,local,sudo,cd,execute
+from fabric.api import run,env,local,cd,settings
 import uuid
 import zipfile
 import tarfile
 import json
 import os
+import jwt
+
 
 env.hosts=['127.0.0.1']
-print(os.getenv('USER'), 'user')
-print(os.getenv('PASSWORD'), 'password')
 env.user = os.getenv("USER")
 env.password = os.getenv("PASSWORD")
 
@@ -61,11 +61,13 @@ def codebase_setup(app):
                 if repo_link.find("git@github.com:") == -1:
                     sp = repo_link.split("github.com")
                     sp = sp[1].strip("/")
-                    print(sp)
                     repo_link = "git@github.com:" + sp + ".git"
                     print(repo_link)
-                run('git clone {} .'.format(repo_link))
-
+                with settings(warn_only=True):
+                    r = run('git clone {} .'.format(repo_link))
+                    if r.failed:
+                        local("rm -rf static")
+                        return ["An error occured while cloning", 400] 
             except:
                 local("rm -rf static")
                 return ["An error occured while cloning", 400]
@@ -79,7 +81,7 @@ def codebase_setup(app):
         try:
             with cd(root_folder):
                 username = app.get("hubUsername")
-                docker_file_path = app.get("dockerfilePath")
+                docker_file_path = app.get("dockerfilePath").strip().lstrip("/")
                 app_name = app.get("appName")
                 app_tag = app.get("appTag")
                 image_name = f"{username}/{app_name}"
@@ -88,12 +90,12 @@ def codebase_setup(app):
                     docker_file_path = "."
                 
                 if app.get('dockerfilePresent') == 'yes':
-                    try:
-                        run("docker build -t {}:{} {}".format(image_name, app_tag, docker_file_path))
-                    except:
-                        local("rm -rf static")
-                        return ["An error occured while building your image from the dockerfile you provided,\
-    check the dockerfile template and the path provided and ensure no error", 400]
+                    with settings(warn_only=True):
+                        r = run("docker build -t {}:{} {}".format(image_name, app_tag, docker_file_path))
+                        if r.failed:    
+                            local("rm -rf static")
+                            return ["An error occured while building your image from the dockerfile you provided,\
+        check the dockerfile template and the path provided and ensure no error", 400]
                 else:
                     custom_dockerfile = app.get("customDockerfile").strip()
                     compose_dockerfile = ""
@@ -126,50 +128,54 @@ def codebase_setup(app):
                             compose_dockerfile += f"{json.dumps(stripped_cmd_array)}\n"
                         for key, port in docker_ports.items():
                             compose_dockerfile += f"EXPOSE {port}\n"
-                        print(compose_dockerfile)
                     try:
                         with open(f"{current_path}/Dockerfile",  "w+") as f:
                             f.write(compose_dockerfile)
                     except:
                         local('rm -rf static')
                         return ["An error occured while saving the dockerfile provided", 400]
-                    try:
-                        password = app.get("hubPassword")
-                        # Building the docker image
-                        run("docker build -t {}:{} .".format(image_name, app_tag))
-                    except:
-                        local('rm -rf static')
-                        return ["An error occured while trying to build image perhaps something went wrong with the dockerfile options you provided", 400]
-                    import jwt
-                    secret_name = f"{str(uuid.uuid4())}{session_id}"
-                    # secret = str(uuid.uuid4()) + "-" + session_id
-                    secret = str(uuid.uuid5(uuid.UUID(session_id), session_id))
-                    encoded_password = jwt.encode({"password": password}, secret, algorithm="HS256")
-                    print(encoded_password)
-                    jwt_dict = {
-                        "enc": encoded_password,
-                        "secret": secret
-                    }
-                    try:
-                        # authenticating into docker hub cli
-                        with open(f"/tmp/{secret_name}", "w") as f:
-                            json.dump(jwt_dict, f)
-                        run(f"chmod 600 /tmp/{secret_name}")
-                        run("cat /tmp/{} | {}/secret_parser.py | docker login -u {} --password-stdin".format(secret_name, current_directory, username))
-                        run(f"rm /tmp/{secret_name}")
-                    except Exception as e:
-                        local('rm -rf static')
-                        local(f'docker rmi -f {image_name}:{app_tag}')
-                        return ["An error occured while trying to login to your dockerhub account", 400]
-                    try:
-                        run("docker push {}:{}".format(image_name, app_tag))
-                    except:
+                    # Building the docker image after composing the dockerfile from the options provided
+                    with settings(warn_only = True):
+                        r = run("docker build -t {}:{} .".format(image_name, app_tag))
+                        if r.failed:
+                            local('rm -rf static')
+                            return ["An error occured while trying to build image perhaps something went wrong with the dockerfile options you provided", 400]
+                
+                # Getting the password and encrypting it
+                password = app.get("hubPassword")
+                secret_name = f"{str(uuid.uuid4())}{session_id}"
+                # secret = str(uuid.uuid4()) + "-" + session_id
+                secret = str(uuid.uuid5(uuid.UUID(session_id), session_id))
+                encoded_password = jwt.encode({"password": password}, secret, algorithm="HS256")
+                jwt_dict = {
+                    "enc": encoded_password,
+                    "secret": secret
+                }
+                try:
+                    # authenticating into docker hub cli
+                    with open(f"/tmp/{secret_name}", "w") as f:
+                        json.dump(jwt_dict, f)
+                    run(f"chmod 600 /tmp/{secret_name}")
+                    with settings(warn_only=True):
+                        r = run("cat /tmp/{} | {}/secret_parser.py | docker login -u {} --password-stdin".format(secret_name, current_directory, username))
+                        if r.failed:
+                            local(f"rm -f /tmp/{secret_name}")
+                            local('rm -rf static')
+                            local(f'docker rmi -f {image_name}:{app_tag}')
+                            return ["An error occured while trying to login to your dockerhub account", 400]
+                    run(f"rm /tmp/{secret_name}")
+                except Exception as e:
+                    local('rm -rf static')
+                    local(f'docker rmi -f {image_name}:{app_tag}')
+                    return ["An error occured", 400]
+                
+                with settings(warn_only=True):
+                    r = run("docker push {}:{}".format(image_name, app_tag))
+                    if r.failed:
                         local('rm -rf static')
                         local(f'docker rmi -f {image_name}:{app_tag}')
                         local("rm ~/.docker/config.json")
-                        return ["Something went wrong while trying to push to dockerhub", 400]
-
-                                                
+                        return ["Something went wrong while trying to push to dockerhub", 400]                                             
         except Exception as e:
             local('rm -rf static')
             return ["Invalid root path provided", 400]
